Add hardcoded archive generation example to test 1
This state is obviously not ready - but it provides an example of how
we can use the llvm ar that zig ships with as a reference black-box
program to test again - using it to autmatically generate our test
cases.

This exmaple is hardcoded - the next step will be to pull it out to a
function & then to use the zig compiler to generate the libraries that
we archive automatically so we no longer need binary files to be checked
in.