Add hardcoded archive generation test
For now this is hardcoded - will need to split it out into functions so
we can test more archive variations. But it's a good first step for
ensuring we can test our archive generation into addition to our
archive parsing. It actually caught a couple of bugs (places where we
differed from llvm) which is nice!