Pass through working directory as argument
Instead of assuming it's std.fs.cwd() everywhere...

This change was needed because it was turning out to be quite tough to
get the tests working properly otherwise (even changing directories back
and forth to the original still made tests conistently fail to
initialise).

However - considering that an intended use case for zar is as a library,
this change is probably still worthwhile doing.