Change zld integration to support latest zig
The old version of zld zar supported didn't support the latest version
of zig for macos object parsing. In order to improve our workflow around
this - this change uses zld as a submodule instead of ad-hoc-ly
embedding it into this project locally. This should hopefully make
upadting zar easier, whilst keeping track of customisations through git
so they can be properly cleaned up later.

This change only stages this support in, the "old" zar still exists and
is used for everything except macos object parsing. The goal is to clean
this up ASAP now that this works for that one case and the tests pass
there.