Make argument handling a lot more flexible
Previously, zar would require arguments to be passed in a specific
order, which was determined to be the correct way to do things based on
the documentation listing them a such.

However, after playing around with llvm ar - in practise we should be
able to deal with modifiers provided (almost) anywhere as an argument.

This change implements that support, making the way we deal with
modifiers a lot more flexible, matching the observed behaviour of llvm
ar more closely in that regard.

However - further testing, and potentially fuzz testing will be needed
to ensure this implementation is actually robust enough to deal with all
the required test cases.