Start MRI parser integration
Accept the -M option and create an MRI parser from it, the
implementation of which is in the middle of a refactor. This is a safe
place for an initial checkin with clear places marked as TODOs.

Before marging the parser will need to be completed and tests for it
will need to be written.