src/main.zig: refactored argument parsing for robustness
zar previously used a while loop and an explicit, manually
incremented, index variable to iterate over its arguments.
This does not clearly convey intent.

The while loop has been replaced with a labelled for, to
control iteration, and a labelled switch, to dispatch an explicit
parser state.

Hopefully this makes the flow of control and data through
the parser a bit clearer.