First pass TestSequence idea implementation
This is kind of hacked together as "Test Sequence Example Test", the
idea being this is an example of how a test sequence might be contructed
for testing zar.

Currently, the existing tests still exist, but the idea would be to
replace them with this setup completely (allowing for the code to be
majorly simplified).

This will allow us to more easily test more features in the future.